glassware.nmr classes' `molecule_atoms` setters modified. Minor style changes.
'molecule_atoms' setter can now take list or list of lists as parameter, but
only one list of atoms is stored. It works this way to stay consistent with
take_atoms and drop_atoms functions from datawork module as well as with
Molecules.arrayed api. This mechanism and functions should change in the future
to support list of not identical molecules.
Appropriate docstrings were updated and unit tests created.

Properties dedclaration changed to decorator protocol.