Added parameter `hard_bound` to datawork.geometry.energy_windows().
This flag controls how window should behave close to borders. With hard bound (True) it will move to the other side of border as soon, as it is reached. With soft bound (False) it will "squeze" when pressed against the border, producing subsequences of the first view that includes border value. True by default, as it feels more natural to not have subsequences as consecutive windows. "Soft" mode left in, because it allows to easily produce a window for each value in the original array.