Removed file extension from conformer identifier.
`.extract_iter()` now returns only a stem of a filename (without file extension) as a conformer identifier. This change was made to better separate concepts of conformer and file with conformer's data. It also enables to connect files with different extension (i.e. .out and .log) but the same name with the same conformer.
`.extract()` uses those identifiers as keys in created dict.
Updated affected tests.