Changed interfaces containing "bars" in name.
"Bars" is an in-house name and is not applicable for publication, as wider public may not recognise what it as a name for spectral activity values. List of changes:
[...]Writer.bars(..., bars) -> .spectral_data(..., data)
datawork.intensities.default_spectra_bars -> .DEFAULT_ACTIVITIES
Tesliper.export_bars() -> .export_spectral_data()
Tesliper.bars -> .activities
References in docstrings, strings, and comments.